Soporte temprano de estructuras
git-svn-id: file:///media/disk/repository/release/trunk@195 17f328a0-2faa-4260-a3e8-86cdd9e0d718

+++ test/simple/struct/struct.c
@@ -24,7 +24,7 @@ struct Estructura3 {
 };
 
 struct Estructura {
-	int entero1;
+	double entero1;
 	int entero2;
 	struct Estructura3 estructura3;
 };
